reorg panther_config_defaults
