+++ global_helpers/panther_base_helpers.py
@@ -10,8 +10,6 @@ from functools import reduce
 from ipaddress import ip_address, ip_network
 from typing import Any, List, Optional, Sequence, Union
 
-from panther_config import config
-
 # # # # # # # # # # # # # #
 #       Exceptions        #
 # # # # # # # # # # # # # #
@@ -20,59 +18,6 @@ from panther_config import config
 class PantherUnexpectedAlert(Exception):
     """Error returned when a Panther detection encounters an unexpected situation."""
 
-
-# # # # # # # # # # # # # #
-#   Compliance Helpers    #
-# # # # # # # # # # # # # #
-
-# Expects a map with a Key 'Tags' that maps to a map of key/value string pairs, or None if no
-# tags are present.
-# All Panther defined resources meet this requirement.
-CDE_TAG_KEY = "environment"
-CDE_TAG_VALUE = "pci"
-
-
-# Defaults to True to assume something is in scope if it is not tagged
-def in_pci_scope_tags(resource):
-    if resource.get("Tags") is None:
-        return True
-    return resource["Tags"].get(CDE_TAG_KEY) == CDE_TAG_VALUE
-
-
-PCI_NETWORKS = config.PCI_NETWORKS
-
-
-# Expects a string in cidr notation (e.g. '10.0.0.0/24') indicating the ip range being checked
-# Returns True if any ip in the range is marked as in scope
-def is_pci_scope_cidr(ip_range):
-    return any(ip_network(ip_range).overlaps(pci_network) for pci_network in PCI_NETWORKS)
-
-
-DMZ_NETWORKS = config.DMZ_NETWORKS
-
-
-# Expects a string in cidr notation (e.g. '10.0.0.0/24') indicating the ip range being checked
-# Returns True if any ip in the range is marked as DMZ space.
-def is_dmz_cidr(ip_range):
-    """This function determines whether a given IP range is within the defined DMZ IP range."""
-    return any(ip_network(ip_range).overlaps(dmz_network) for dmz_network in DMZ_NETWORKS)
-
-
-# Defaults to False to assume something is not a DMZ if it is not tagged
-def is_dmz_tags(resource, dmz_tags):
-    """This function determines whether a given resource is tagged as existing in a DMZ."""
-    if resource["Tags"] is None:
-        return False
-    for key, value in dmz_tags:
-        if resource["Tags"].get(key) == value:
-            return True
-    return False
-
-
-# Function variables here so that implementation details of these functions can be changed without
-# having to rename the function in all locations its used, or having an outdated name on the actual
-# function being used, etc.
-IN_PCI_SCOPE = in_pci_scope_tags
 
 # # # # # # # # # # # # # #
 #      GSuite Helpers     #
