+++ policies/aws_load_balancer_policies/aws_elbv2_load_balancer_has_ssl_policy.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 
 def policy(resource):
