+++ policies/aws_rds_policies/aws_rds_instance_backup_retention_acceptable.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 MAX_RETENTION_DAYS = 180
 MIN_RETENTION_DAYS = 7
