+++ policies/aws_redshift_policies/aws_redshift_cluster_snapshot_retention_acceptable.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 # Retention period in days
 MIN_RETENTION_DAYS = 3
