+++ policies/aws_s3_policies/aws_s3_bucket_object_lock_configured.py
@@ -1,4 +1,5 @@
-from panther_base_helpers import IN_PCI_SCOPE, deep_get
+from panther_base_helpers import deep_get
+from panther_config_defaults import IN_PCI_SCOPE
 
 RETENTION_PERIOD_DAYS = 365
