+++ policies/aws_vpc_policies/aws_network_acl_restricts_insecure_protocols.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 # This is a list of default ports for insecure protocols. As AWS Network ACLs and Security Groups
 # are not application layer aware, this is the closest approximation that can be made to blocking
