+++ policies/aws_vpc_policies/aws_network_acl_restricts_outbound_traffic.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 # This is generic policy that checks outbound traffic rules on a Network ACL.
 # It is recommended you add additional logic for your own use cases.
