+++ policies/aws_vpc_policies/aws_security_group_restricts_inter_security_group_traffic.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 
 def policy(resource):
