+++ policies/aws_vpc_policies/aws_security_group_restricts_outbound_traffic.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 # This is a generic policy that checks outbound permissions on Security Groups.
 # You may wish to add additional logic specific to your use case.
