+++ policies/aws_vpc_policies/aws_security_group_restricts_traffic_leaving_cde.py
@@ -1,6 +1,6 @@
 from ipaddress import ip_network
 
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 
 def policy(resource):
