+++ policies/aws_vpc_policies/aws_security_group_tightly_restricts_outbound_traffic.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 
 MAX_PORTS_PER_PERMISSION = 10
 RESTRICTED_PORTS = [
