+++ policies/aws_vpc_policies/aws_vpc_default_network_acl_restricts_all_traffic.py
@@ -1,4 +1,4 @@
-from panther_base_helpers import IN_PCI_SCOPE
+from panther_config_defaults import IN_PCI_SCOPE
 from panther_oss_helpers import BadLookup, resource_lookup
