+++ policies/aws_waf_policies/aws_waf_has_xss_predicate.py
@@ -1,4 +1,5 @@
-from panther_base_helpers import IN_PCI_SCOPE, deep_get
+from panther_base_helpers import deep_get
+from panther_config_defaults import IN_PCI_SCOPE
 
 
 def policy(resource):
